build(webpack): Added stylelint and renamed scss source directory
Added stylelint, stylelint-scss and stylelint-webpack-plugin as dependencies. Now the scss files
gets linted and automatically fixed (not for all rules) when there are any errors. Also renamed the
directory for the SCSS source files to "scss" from "sass". Fixed import in the Electron renderer
process. Also applied the stylelint fixes to the source files