build(webpack): Removed tslint completly from webpack
Instead we added chokidar-cli as dependency and provide scripts for the linting process. Also
refactored the webpack configuration files. The common config now uses the 'mini-css-extract-plugin'
and the 'html-webpack-plugin' so we don't need to specify them in the development and production
config. It also provides now sourcemaps when webpack does not runs in production mode. The pattern
for SASS files now only matches '.scss' files and not '.sass' and '.scss' anymore.